Overhaul for chest display
Chest types are now included in an array so that multiple types can now be entered for one chest and will be filtered correctly if any of the tags are selected. Updated the way currency displays so if the item has a set sell type with no range, it will simply display the item value rather than a range of 0. Updated the placement of text slightly.